Refactor loess structure and its setup/init
When the loess structure is made up of pointers to its sub-structures
(inputs, model, ...), the relation is clearer in cython. The
sub-structures can be fully created when `__init__` completes. The
doc-string is accurate.

Again, memory is always allocated for pseudovalues. Should it be
necessary to change this, then a NULL-malloc-free-NULL juggle can
be added.