Refactor memory allocation and __dealloc__
Since the `loess` sub-structs are pointers, and in cython they
initialized in `__cinit__`, the memory deallocation was also
split up. The structure of the cython and the C look more
similar.

Between this, and the previous commit a bunch of segfaults that
did not affect the current tests have been cleared up.

The array creating helper functions have also been updated to
handle 2D arrays.