Fix corrupted array views
`smisc.loess` uses object components for inputs, initialization
control, model, output and prediction. These components are
windows to the C & fortran implementation code, and through
them we can access the memory buffers used in the
implementation. The problem was when a buffer was wrapped
to create a numpy array and handed to the user, the array
was only gauranteed to be valid if the component managing
it continued to exist (i.e. it was not gabagge collected).

The solution is not only pass copies of the buffers to the
user API.

fixes #8